# docs/api.yaml
basePath: /
host: localhost:3344
info:
  description: Backend of repo https://github.com/vidorg/vid_vue
  license:
    name: MIT
    url: https://github.com/vidorg/vid_backend/blob/master/LICENSE
  termsOfService: https://github.com/vidorg
  title: vid backend
  version: '1.1'
paths:
...
  /v1/auth/:
    get:
      consumes:
      - multipart/form-data
      description: 根据认证令牌，查看当前登录用户
      operationId: -v1-auth-get
      parameters:
      - &id001
        description: 用户登录令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"uid\": 10,\n        \"username\": \"aoihosizora\"\
            ,\n        \"sex\": \"male\",\n        \"profile\": \"Demo profile\",\n\
            \        \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n        \"birth_time\": \"2001-02-03\",\n        \"authority\": \"\
            normal\",\n        \"phone_number\": \"13512345678\"\n    }\n}\n```"
        '401': &id002
          description: '"authorization failed" / "token has expired"'
      security:
      - basicAuth: ''
      summary: 查看当前登录用户
      tags:
      - Authorization
  /v1/auth/login:
    post:
      consumes:
      - multipart/form-data
      description: 用户登录
      operationId: -v1-auth-login-post
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 用户密码
        in: formData
        name: password
        required: true
        type: string
      - description: 登录有效期，默认为七天
        in: formData
        name: expire
        required: false
        type: integer
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"user\": {\n            \"uid\": 10,\n  \
            \          \"username\": \"aoihosizora\",\n            \"sex\": \"male\"\
            ,\n            \"profile\": \"Demo profile\",\n            \"avatar_url\"\
            : \"http://localhost:3344/raw/image/default/avatar.jpg\",\n          \
            \  \"birth_time\": \"2001-02-03\",\n            \"authority\": \"normal\"\
            ,\n            \"phone_number\": \"13512345678\"\n        },\n       \
            \ \"token\": \"Bearer xxx\",\n        \"expire\": 604800\n    }\n}\n```"
        '400':
          description: '"request param error"'
        '401':
          description: '"password error"'
        '404':
          description: '"user not found"'
        '500':
          description: '"login failed"'
      summary: 登录
      tags:
      - Authorization
  /v1/auth/password:
    put:
      consumes:
      - multipart/form-data
      description: 用户修改密码
      operationId: -v1-auth-password-put
      parameters:
      - *id001
      - description: '用户密码

          *Minimum length* : 8

          *Maximum length* : 30'
        in: formData
        maxLength: 30
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            \n}\n```"
        '400':
          description: '"request param error" / "request format error"'
        '401': *id002
        '404':
          description: '"user not found"'
        '500':
          description: '"update password failed"'
      security:
      - basicAuth: ''
      summary: 修改密码
      tags:
      - Authorization
  /v1/auth/register:
    post:
      consumes:
      - multipart/form-data
      description: 用户注册
      operationId: -v1-auth-register-post
      parameters:
      - description: '用户名

          *Minimum length* : 5

          *Maximum length* : 30'
        in: formData
        maxLength: 30
        minLength: 5
        name: username
        required: true
        type: string
      - description: '用户密码

          *Minimum length* : 8

          *Maximum length* : 30'
        in: formData
        maxLength: 30
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"code\": 200,\n    \"message\": \"success\"\
            ,\n    \"data\": {\n        \"uid\": 10,\n        \"username\": \"aoihosizora\"\
            ,\n        \"sex\": \"male\",\n        \"profile\": \"Demo profile\",\n\
            \        \"avatar_url\": \"http://localhost:3344/raw/image/default/avatar.jpg\"\
            ,\n        \"birth_time\": \"2001-02-03\",\n        \"authority\": \"\
            normal\",\n        \"phone_number\": \"13512345678\"\n    }\n}\n```"
        '400':
          description: '"request param error" / "request format error"'
        '500':
          description: '"username has been used" / "register failed"'
      summary: 注册
      tags:
      - Authorization
  /v1/raw/image/{uid}/{filename}:
    get:
      consumes:
      - multipart/form-data
      description: 获取用户头像图片以及视频封面
      operationId: -v1-raw-image-{uid}-{filename}-get
      parameters:
      - description: 用户id，或者default
        in: path
        name: uid
        required: true
        type: string
      - description: 图片文件名，jpg后缀名
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        '200':
          description: "```json\n{\n    \"Content-Type\": \"image/jpeg\"\n}\n```"
        '400':
          description: '"request param error"'
        '404':
          description: '"image not found"'
      summary: 获取图片
      tags:
      - Raw
...
swagger: '2.0'